requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
chromadb
langchain
langchain-community
langchain-google-genai
